pyproject.toml
[project]
name = "body-composition"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.13",
    "matplotlib>=3.10.0",
    "notebook>=7.3.2",
    "pandas>=2.2.3",
    "plotly>=6.1.2",
    "polars>=1.30.0",
    "pyarrow>=20.0.0",
    "pymongo==4.13.2",
    "python-dotenv>=1.1.0",
    "scikit-learn>=1.6.1",
    "seaborn>=0.13.2",
    "uvicorn>=0.34.3",
]
